Moving check_unique test to util tests
